✨ feat: 뒤로가기 눌렀을 때 모달 닫히도록 구현 #8
